Add --install_profile option
By default an installation of lux is a minimal standalone system,
where InstallDir is the directory where the system will be installed.

  --install=<InstallDir>

Now the installation can be customized by using the following options:

  --install_app=<ErlangApp>

is used to add more Erlang applications, such as the debugger and wx if
you want to debug lux itself with lux -D. But in order to include more
executables, include files etc, the Reltool profile must be changed from
standalone (being the default) to development. This can be done with:

  --install_profile=<ReltoolProfile>

For example, the following command will install lux (and the apps that
lux needs), plus mnesia and snmp. Further, all the startup scripts (such
as erl, erlc, epmd etc) are included.

  lux \
    --install=/tmp/lux \
    --install_profile=development \
    --install_app=mnesia \
    --install_app=snmp

With this it is possible to make use of the Erlang system in the Lux
installation from lux scripts. Like this:

  [shell erl]

    !erl
    ?Erlang/OTP
    ?1>

    !code:get_path().
    """???
    [".","/tmp/lux/lib/kernel-2.16.4.ez/kernel-2.16.4/ebin",
     "/tmp/lux/lib/stdlib-1.19.4.ez/stdlib-1.19.4/ebin",
     "/tmp/lux/lib/sasl-2.3.4.ez/sasl-2.3.4/ebin",
     "/tmp/lux/lib/lux-1.6.ez/lux-1.6/ebin",
     "/tmp/lux/lib/compiler-4.9.4.ez/compiler-4.9.4/ebin"]
    2>
    """