Introduce newshell command mode
By default the [shell] command is used both for starting new shells as
well as changing the active shell.

In --newshell mode, shells can only be created with the [newshell]
command and the [shell] command becomes restricted to only change the
active shell.

The --newshell mode can explicitely be activated as a configuration
parameter. Or implicitely by using the [newshell] command.