Handle race condition when shell is terminated
When the shell was terminated, the pending output was sometimes not
matched the final time. It was very timing dependent. This situation
manifested itself as the error message “The command must be executed
in context of a shell”. For the record, that error message could
originate from other contexts as well, not addressed by this fix.
The case that has been fixed was "No more output to match".